+++ docker-compose.yml
@@ -2,7 +2,7 @@ version: '3.1'
 services:
   stats-server:
     build: ewm-stats-service/service
-    container_name: stats-server
+    container_name: stats-service
     depends_on:
       - stats-db
     environment:
@@ -27,8 +27,9 @@ services:
       - ewm_backend
 
   ewm-service:
+    image: ewm-main
     build: ewm-main-service
-    container_name: emw-service
+    container_name: main-service
     ports:
       - "8080:8080"
     depends_on:
@@ -38,20 +39,20 @@ services:
       - /etc/localtime:/etc/localtime:ro
     environment:
       - STATS_SERVER_URL=http://stats-server:9090
-      - SPRING_DATASOURCE_URL=jdbc:postgresql://ewm-db:5432/ewm_main
+      - SPRING_DATASOURCE_URL=jdbc:postgresql://ewm-db:5432/ewm_db
     networks:
       - ewm_backend
 
   ewm-db:
     image: postgres:14-alpine
-    container_name: main-db
+    container_name: ewm_db
     ports:
       - "6542:5432"
     volumes:
       - /etc/timezone:/etc/timezone:ro
       - /etc/localtime:/etc/localtime:ro
     environment:
-      - POSTGRES_DB=emw-db
+      - POSTGRES_DB=ewm_db
     env_file:
       - "pass.env"
     networks:
