+++ ewm-main-service/pom.xml
@@ -20,13 +20,6 @@
 
 
     <dependencies>
-<!--        Warning:(23, 9)  Provides transitive vulnerable dependency org.springframework.boot:spring-boot:2.7.9 CVE-2023-34055 6.5 Uncontrolled Resource Consumption vulnerability with Medium severity found  Results powered by Checkmarx(c) -->
-<!--        Warning:(23, 9)  Provides transitive vulnerable dependency org.springframework.boot:spring-boot-autoconfigure:2.7.9 CVE-2023-20883 7.5 Uncontrolled Resource Consumption vulnerability  Results powered by Checkmarx(c)-->
-<!--        Warning:(23, 9)  Provides transitive vulnerable dependency ch.qos.logback:logback-classic:1.2.11 CVE-2023-6378 7.5 Deserialization of Untrusted Data vulnerability with High severity found  Results powered by Checkmarx(c)-->
-<!--        Warning:(23, 9)  Provides transitive vulnerable dependency ch.qos.logback:logback-core:1.2.11 CVE-2023-6378 7.5 Deserialization of Untrusted Data vulnerability with High severity found  Results powered by Checkmarx(c)-->
-<!--        Warning:(23, 9)  Provides transitive vulnerable dependency org.yaml:snakeyaml:1.30 CVE-2022-25857 7.5 Uncontrolled Resource Consumption vulnerability CVE-2022-38752 6.5 Out-of-bounds Write vulnerability CVE-2022-38749 6.5 Out-of-bounds Write vulnerability CVE-2022-38750 5.5 Out-of-bounds Write vulnerability CVE-2022-38751 6.5 Out-of-bounds Write vulnerability CVE-2022-41854 6.5 Out-of-bounds Write vulnerability CVE-2022-1471 9.8 Deserialization of Untrusted Data vulnerability  Results powered by Checkmarx(c)-->
-<!--        Warning:(23, 9)  Provides transitive vulnerable dependency org.apache.tomcat.embed:tomcat-embed-core:9.0.71 CVE-2023-24998 7.5 Allocation of Resources Without Limits or Throttling vulnerability CVE-2023-28708 4.3 Unprotected Transport of Credentials vulnerability CVE-2023-28709 7.5 Off-by-one Error vulnerability with High severity found CVE-2023-41080 6.1 URL Redirection to Untrusted Site ("Open Redirect") vulnerability with Medium severity found CVE-2023-45648 5.3 Improper Input Validation vulnerability with Medium severity found CVE-2023-42795 5.3 Incomplete Cleanup vulnerability with Medium severity found CVE-2023-46589 7.5 Inconsistent Interpretation of HTTP Requests ("HTTP Request Smuggling") vulnerability with High severity found CVE-2023-42794 5.9 Incomplete Cleanup vulnerability with Medium severity found  Results powered by Checkmarx(c)-->
-<!--        Warning:(23, 9)  Provides transitive vulnerable dependency org.springframework:spring-webmvc:5.3.25 CVE-2016-1000027 9.8 Deserialization of Untrusted Data vulnerability CVE-2023-20860 7.5 Improper Neutralization of Wildcards or Matching Symbols vulnerability  Results powered by Checkmarx(c)-->
         <dependency>
             <groupId>ru.practicum</groupId>
             <artifactId>client</artifactId>
@@ -40,7 +33,6 @@
             <version>2.0.2</version>
         </dependency>
 
-<!--        Warning:(36, 9)  Dependency com.fasterxml.jackson.core:jackson-databind:2.13.3 is vulnerable CVE-2022-42003 7.5 Deserialization of Untrusted Data vulnerability CVE-2022-42004 7.5 Deserialization of Untrusted Data vulnerability  Results powered by Checkmarx(c)-->
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
@@ -56,11 +48,14 @@
             <artifactId>jakarta.persistence-api</artifactId>
         </dependency>
 
-<!--        Warning:(51, 9)  Dependency org.springframework:spring-web:5.3.25 is vulnerable CVE-2016-1000027 9.8 Deserialization of Untrusted Data vulnerability  Results powered by Checkmarx(c)-->
-<!--        Warning:(56, 9)  Provides transitive vulnerable dependency org.springframework:spring-expression:5.3.25 CVE-2023-20863 6.5 Improper Neutralization of Special Elements used in an Expression Language Statement ('Expression Language Injection') vulnerability CVE-2023-20861 6.5 Allocation of Resources Without Limits or Throttling vulnerability with Medium severity found  Results powered by Checkmarx(c)-->
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-web</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-actuator</artifactId>
         </dependency>
 
         <dependency>
@@ -85,4 +80,13 @@
             <version>3.12.0</version>
         </dependency>
     </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+            </plugin>
+        </plugins>
+    </build>
 </project>
