+++ ewm-main-service/src/main/java/ru/practicum/category/api/service/CategoryServiceImpl.java
@@ -24,8 +24,8 @@ import static ru.practicum.constants.Constants.checkPageable;
 @Slf4j
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
-    private final CategoryRepository categoryRepository;
-    private final EventRepository eventRepository;
+    private CategoryRepository categoryRepository;
+    private EventRepository eventRepository;
 
     @Override
     public CategoryDto create(NewCategoryDto newDto) {
