+++ ewm-main-service/src/main/java/ru/practicum/user/api/service/UserServiceImpl.java
@@ -1,0 +1,57 @@
+package ru.practicum.user.api.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import ru.practicum.exception.NotFoundException;
+import ru.practicum.user.api.dto.NewUserDto;
+import ru.practicum.user.api.dto.UserDto;
+import ru.practicum.user.api.mapper.UserMapper;
+import ru.practicum.user.api.repository.UserRepository;
+import ru.practicum.user.entity.User;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static java.util.stream.Collectors.toList;
+import static ru.practicum.constants.Constants.USER_NOT_EXISTS;
+import static ru.practicum.constants.Constants.checkPageable;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class UserServiceImpl implements UserService {
+    private UserRepository repository;
+
+    @Override
+    public UserDto create(NewUserDto newDto) {
+        log.debug("[i] create user");
+        User user = UserMapper.INSTANCE.toEntity(newDto);
+
+        return UserMapper.INSTANCE.toDto(repository.save(user));
+    }
+
+    @Override
+    public List<UserDto> getAll(List<Long> ids, Integer from, Integer size) {
+        log.debug("[i] get users");
+
+        Pageable pageable = checkPageable(from, size, null);
+        List<User> allUsersByIds = repository.findAllById(ids, pageable)
+                .orElse(new ArrayList<>());
+        return allUsersByIds.stream()
+                .map(UserMapper.INSTANCE::toDto)
+                .collect(toList());
+    }
+
+
+    @Override
+    public void remove(long id) {
+        log.debug("[i] delete user ID:{}", id);
+        try {
+            repository.deleteById(id);
+        } catch (RuntimeException e) {
+            throw new NotFoundException(String.format(USER_NOT_EXISTS, id));
+        }
+    }
+}
