+++ ewm-main-service/src/main/resources/application.properties
@@ -1,0 +1,10 @@
+server.port=8080
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL10Dialect
+spring.jpa.properties.hibernate.format_sql=true
+spring.sql.init.mode=always
+#---
+spring.datasource.driverClassName=org.postgresql.Driver
+spring.datasource.url=jdbc:postgresql://localhost:6542/ewm-db
+spring.datasource.username=root
+spring.datasource.password=root
